requirements.txt
# Polymarket Arbitraj — Bağımlılıklar
# Kurulum: python3 -m venv .venv && .venv/bin/pip install -r requirements.txt

# Polymarket CLOB
py-clob-client>=0.18.0
python-dotenv>=1.0.0

# HTTP / WebSocket
requests>=2.28.0
websocket-client>=1.6.0

# Veri / optimizasyon
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# LLM (Google Gemini)
google-generativeai>=0.8.0

# Config
pyyaml>=6.0

# Polygon / Alchemy (eth_getLogs)
web3>=6.0.0

# Dashboard
streamlit>=1.28.0

# Alert e-posta (Resend)
resend>=2.0.0

# Opsiyonel: IP çözücü (lisans gerekir)
# gurobipy>=10.0.0
